requirements.txt
Django==1.6.5
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.5.1
Pygments==1.6
South==1.0
Sphinx==1.2.2
django-appconf==0.6
django-braces==1.4.0
django-classy-tags==0.5.1
django-crispy-forms==1.4.0
django-debug-toolbar==1.2.1
django-imagekit==3.2.2
django-modeltranslation==0.7.3
django-sekizai==0.7
docutils==0.12
gnureadline==6.3.3
ipython==2.1.0
mock==1.0.1
model-mommy==1.2.1
pilkit==1.1.9
six==1.7.3
sqlparse==0.1.11
wsgiref==0.1.2
